Rename `renderer` index files
