+++ src/shared/services/papi.service.ts
@@ -14,13 +14,9 @@ import { isRenderer } from '@shared/utils/internal-util';
 import internetService from '@shared/services/internet.service';
 
 // TODO: Fix these to use NormalModuleReplacementPlugin or something https://webpack.js.org/plugins/normal-module-replacement-plugin/
-const papiComponents = isRenderer()
-  ? require('@renderer/components/papi-components/papi-components').default
-  : {};
-const papiContext = isRenderer()
-  ? require('@renderer/context/papi-context/papi-context').default
-  : {};
-const papiHooks = isRenderer() ? require('@renderer/hooks/papi-hooks/papi-hooks').default : {};
+const papiComponents = isRenderer() ? require('@renderer/components/papi-components').default : {};
+const papiContext = isRenderer() ? require('@renderer/context/papi-context').default : {};
+const papiHooks = isRenderer() ? require('@renderer/hooks/papi-hooks').default : {};
 
 export default {
   // Classes
